# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Simar's personal dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /progress/{category}:
    get:
      operationId: getCategoryProgress
      tags: [progress]
      summary: Get progress for a specific category
      parameters:
        - name: category
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Category progress details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryProgress"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /roadmap-items:
    get:
      operationId: getRoadmapItems
      tags: [roadmap]
      summary: List roadmap items, optionally filtered by category
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of roadmap items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RoadmapItem"
    post:
      operationId: createRoadmapItem
      tags: [roadmap]
      summary: Create a new roadmap item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoadmapItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoadmapItem"
  /roadmap-items/{id}:
    patch:
      operationId: updateRoadmapItem
      tags: [roadmap]
      summary: Update a roadmap item (toggle completion, rename)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoadmapItemUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoadmapItem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteRoadmapItem
      tags: [roadmap]
      summary: Delete a roadmap item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
  /wins:
    get:
      operationId: getWins
      tags: [wins]
      summary: Get all wins with monthly summary
      responses:
        "200":
          description: Wins data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WinsSummary"
    post:
      operationId: addWin
      tags: [wins]
      summary: Add a new win
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WinInput"
      responses:
        "201":
          description: Win added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Win"
  /wins/{id}:
    delete:
      operationId: deleteWin
      tags: [wins]
      summary: Delete a win
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
...
  /vision-board:
    get:
      operationId: getVisionBoard
      tags: [vision]
      summary: Get all vision board items
      responses:
        "200":
          description: List of vision board items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VisionItem"
    post:
      operationId: createVisionItem
      tags: [vision]
      summary: Add a vision board item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VisionItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VisionItem"
  /vision-board/{id}:
    delete:
      operationId: deleteVisionItem
      tags: [vision]
      summary: Delete a vision board item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
...
  /quotes:
    get:
      operationId: getQuotes
      tags: [quotes]
      summary: List all quotes, optionally filtered by category
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of quotes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Quote"
    post:
      operationId: createQuote
      tags: [quotes]
      summary: Add a quote
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuoteInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quote"
  /quotes/random:
    get:
      operationId: getRandomQuote
      tags: [quotes]
      summary: Get a random quote, optionally filtered by category
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: A random quote
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quote"
  /quotes/{id}:
    delete:
      operationId: deleteQuote
      tags: [quotes]
      summary: Delete a quote
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    CategoryProgress:
      type: object
      required: [category, label, emoji, percentage, completedItems, totalItems]
      properties:
        category:
          type: string
        label:
          type: string
        emoji:
          type: string
        percentage:
          type: number
        completedItems:
          type: number
        totalItems:
          type: number
    RoadmapItem:
      type: object
      required: [id, category, title, completed, order, createdAt]
      properties:
        id:
          type: number
        category:
          type: string
        title:
          type: string
        completed:
          type: boolean
        order:
          type: number
        createdAt:
          type: string
          format: date-time
        notes:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        contactEmail:
          type: ["string", "null"]
        linkUrl:
          type: ["string", "null"]
    RoadmapItemInput:
      type: object
      required: [category, title]
      properties:
        category:
          type: string
        title:
          type: string
        order:
          type: number
    RoadmapItemUpdate:
      type: object
      properties:
        title:
          type: string
        completed:
          type: boolean
        order:
          type: number
        notes:
          type: string
        status:
          type: string
        contactEmail:
          type: string
        linkUrl:
          type: string
    Win:
      type: object
      required: [id, description, createdAt]
      properties:
        id:
          type: number
        description:
          type: string
        createdAt:
          type: string
          format: date-time
    WinInput:
      type: object
      required: [description]
      properties:
        description:
          type: string
    WinsSummary:
      type: object
      required: [wins, totalThisMonth, total]
      properties:
        wins:
          type: array
          items:
            $ref: "#/components/schemas/Win"
        totalThisMonth:
          type: number
        total:
          type: number
...
    VisionItem:
      type: object
      required: [id, type, content, createdAt]
      properties:
        id:
          type: number
        type:
          type: string
          description: "image or quote"
        content:
          type: string
          description: "URL for images, text for quotes"
        caption:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    VisionItemInput:
      type: object
      required: [type, content]
      properties:
        type:
          type: string
        content:
          type: string
        caption:
          type: string
...
    Quote:
      type: object
      required: [id, text, category]
      properties:
        id:
          type: number
        text:
          type: string
        category:
          type: string
        author:
          type: ["string", "null"]
    QuoteInput:
      type: object
      required: [text, category]
      properties:
        text:
          type: string
        category:
          type: string
        author:
          type: string
...