# lib/api-zod/src/generated/types/categoryProgress.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Simar's personal dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface CategoryProgress {
  category: string;
  label: string;
  emoji: string;
  percentage: number;
  completedItems: number;
  totalItems: number;
}