# lib/api-zod/src/generated/types/roadmapItemUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Simar's personal dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface RoadmapItemUpdate {
  title?: string;
  completed?: boolean;
  order?: number;
  notes?: string;
  status?: string;
  contactEmail?: string;
  linkUrl?: string;
}