# lib/api-zod/src/generated/types/visionItemInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Simar's personal dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface VisionItemInput {
  type: string;
  content: string;
  caption?: string;
}