# lib/api-zod/src/generated/types/win.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Simar's personal dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface Win {
  id: number;
  description: string;
  createdAt: Date;
}